modified:   long.cpp 	modified:   run

+++ long.cpp
@@ -3,7 +3,7 @@
 int main() {
     int n=1;
     int repeats=100000000;
-    long pi=1;
+    double pi=1;
     while (n<repeats) {
         pi=pi * 2*n/(2*n-1) * 2*n/(2*n+1);
         n++;
